Move frontend blocks to their specific folder
