+++ assets/js/frontend/components/load-more-button/index.js
