+++ assets/js/frontend/components/read-more/index.js
