+++ assets/js/frontend/components/read-more/utils.js
