+++ assets/js/frontend/hocs/with-component-id.js
