+++ assets/js/hocs/index.js
@@ -1,3 +1,2 @@
-export { default as withComponentId } from './with-component-id';
 export { default as withProduct } from './with-product';
 export { default as withSearchedProducts } from './with-searched-products';
